delete performed

+++ src/pages/MyList.jsx
@@ -3,6 +3,7 @@ import Navbar from "../components/Navbar";
 import Footer from "../components/Footer";
 import useAuth from "../hooks/useAuth";
 import { Helmet } from "react-helmet";
+import Swal from "sweetalert2";
 
 const MyList = () => {
   const { user } = useAuth() || {};
@@ -16,32 +17,47 @@ const MyList = () => {
       });
   }, [user]);
 
-  const handleDelete = id => {
+  const handleDelete = (id) => {
     //make sure he confirms
     fetch(`http://localhost:5000/myList/${user?.email}/${id}`, {
-        method: "DELETE",
-        // headers: {
-        //     'content-type': 'application/json'
-        // },
-        // body: JSON.stringify
+      method: "DELETE",
+      // headers: {
+      //     'content-type': 'application/json'
+      // },
+      // body: JSON.stringify
     })
-    .then(res => res.json())
-    .then(data => {
-        if(data.deletedCount>0){
-            console.log('deleted success!');
-
-            //remove the list item from UI
-            const remainingItem = item.filter(i => i._id !== id);
-            setItem(remainingItem);
+      .then((res) => res.json())
+      .then((data) => {
+        if (data.deletedCount > 0) {
+          Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!",
+          }).then((result) => {
+            if (result.isConfirmed) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+              //remove the list item from UI
+              const remainingItem = item.filter((i) => i._id !== id);
+              setItem(remainingItem);
+            }
+          });
         }
-    })
-  }
+      });
+  };
 
   return (
     <div>
-        <Helmet>
-            <title>Explorex- My List</title>
-        </Helmet>
+      <Helmet>
+        <title>Explorex- My List</title>
+      </Helmet>
       <Navbar></Navbar>
 
       <div className="w-11/12 mx-auto my-10">
@@ -67,8 +83,15 @@ const MyList = () => {
                   <td>{p.country}</td>
                   <td>${p.cost}</td>
                   <td className="flex flex-col gap-3">
-                    <button className="btn btn-accent text-white">Update</button>
-                    <button onClick={() => handleDelete(p._id)} className="btn btn-error text-white">Delete</button>
+                    <button className="btn btn-accent text-white">
+                      Update
+                    </button>
+                    <button
+                      onClick={() => handleDelete(p._id)}
+                      className="btn btn-error text-white"
+                    >
+                      Delete
+                    </button>
                   </td>
                 </tr>
               ))}
